app_with_db/requirements.txt
uWSGI==2.0.21
Flask==2.3.2
sqlalchemy
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
